First commit on the tool_registry branch. Design philosophy:
1) make tools/ a sibling dir to agents/ so we have
core/
|-- agents/
|-- tools/

This seems to make the most sense as end users will want to interact with agents (doing inference) and tools (registering their own and selecting builtins). So while agents has a dependency on tools it does not seem right to make agents a subdir of tools or vice versa.

2) Making tools/ sibling to agents/ creates an import challenge

In order for research_agent.py to import core/tools/registry as `import core.tools.registry`, I added the project root path to the system python path. There's probably a more elegant way to do this

3) core/tools/builtins can house all the builtin tools we'll ever define